Update README, and markup with reStructuredText.
Update the API documentation, since bccCompileBC and bccLoadBinary
are removed.  Update the document about the cache file format.
Since the caching mechanism has been rewritten completely.

Note: Now, we are going to adopt reStructuredText to markup
the document.  reStructuredText is a natural markup language,
which is human-eye-friendly (contrast to html).