Check libbcc build time instead of sha1sum.
Since libbcc.so is larger than 3+Mb, it takes a lot of time
to calculate its sha1 checksum.  In order to shorten the startup
time, we introduce the build time check.  There will be a string
called libbcc_build_time in libbcc.so and the cached file,
which will store the build time of libbcc.so.

In case that you have to upate libbcc.so frequently, and you
want to enable the sha1 checksum of libbcc, then you may change
USE_LIBBCC_SHA1SUM to true.