Add proper build-time configs for VFP/NEON/...
This change fixes the way that libbcc is built. First, it properly sets up
all required libbcc_CFLAGS-specified defines (i.e. ARCH_ARM_HAVE_NEON) to be
passed to the compiler. It also builds the host binary (standalone bcc) to
cross-compile for the target architecture (along with associated features).
This will aid in debugging NEON issues.

Change-Id: If499e944244493ca65ceb691b097bad6b6da94f5