Fix RSExpandForEach pass.
The bug here is that the c_str() storage of a temporary std::string is not
guaranteed to persist beyond the scope of the std::string. In this case, we
were keeping a const char * to freed memory (and using garbage values to
populate our list of functions to keep).

Change-Id: I40712471c05dec1847c08dee1d6db13070c3e6ac