Fix RSExpandForEach pass.
The bug here is that the c_str() storage of a temporary std::string is not
guaranteed to persist beyond the scope of the std::string. In this case, we
were keeping a const char * to freed memory (and using garbage values to
populate our list of functions to keep).

Change-Id: I6e1bd64bbf302c004a34c926b2f7cd643ed98e76