Fix pass-by-value when the return type uses pass-by-reference.
Since we are using a baseline ARM ABI for Renderscript's portable bitcode,
we need to handle the case where a large return value gets passed as a
pointer (it becomes the first argument). This also fixes a small issue
where the first argument could be the return value (i.e. passed by
reference). I also have verified that coerced struct pass-by-value is
now working with this change.

Bug: 7166741

Change-Id: I7c00ff437ad862f9b8b244d87ccbca19e46c5775