ForEachExpand: Construct loop in a separate function
When expanding the loading/unloading of parameter values as well as the
creation of the implicit loop where mangled together. We now separate
them. This reduces code duplication and makes the difference between
the two ways of expanding a kernel more clear.

We also add some documentation and a FIXME that highlights the
surprising use of NUW during code generation.

The only functional change included here is that we switch to using PHI-nodes
instead of loading/storing the loop iterator to memory.

Change-Id: I8c28f3054eb00a73b8b754057625e52140c04405