Remove manually inlined populateLTOPassManager()
Instead of copying the content of
llvm::PassManagerBuilder::populateLTOPassManager() we just call it.
This does not only reduce code duplication, but also ensures that we do not
miss anything. Specifically, the previous code did not run any alias analysis
at all. After this change we now always run tbaa and basicaa. This does not
give any benefits at the moment, but is a prerequisite for upcoming changes to
our alias analysis.

Change-Id: I01b352255e958cbc1c996d934a519f5ab71f9f24