Run verifier on RenderScript .bc modules
Verify the bitcode to abort gracefully on invalid .bc modules instead of
crashing randomly. This patch will not detect all problems with the bitcode
writing, but it is likely to catch attribute breakages such as Bug 9503142. It
also prevents us from crashing on the incorrect debug info found in Bug
1031440.

Change-Id: I2bf626b7aae173dc7238a499b2a0474301197cbe