Ensure that we only compile the proper bitwidth bitcode for an architecture.
Bug: 16031597

Prior to this change, it was possible to compile 32-bit bitcode for a 64-bit
architecture and/or 64-bit bitcode for a 32-bit architecture. This change
enforces that the target architecture's pointer width matches the pointer
width of the individual RS triples that we work with.

Change-Id: Ida65582875fb061911fc6a92cd99454f1ab6c5e6
(cherry picked from commit 10ee6af612d585301c6f2b2f0f04e9bc80b338bb)