Remember to mark functions as materializable (or they are only "declared").
Bug: 18761789

Several relevant changes were made to BitcodeReader that we failed to update
on the last rebase. This results in functions not being properly translated
to the latest bitcode revision (which then causes the compiler to fail).
BitcodeReader::isMaterializable() is dead code, so I removed it as part of
this change, along with a missed 0 -> nullptr update.

Change-Id: Ifa5dfe4b146e1469e937fda0a107bb2c21936071