Fix TBAA by using an additional distinct node for swapping.
Bug: 20069947

We now create a distinct TBAA root node, so that we can easily swap the rest
of the RenderScript TBAA subtree into the simple C/C++ TBAA tree. This change
also reenables TBAA for RenderScript, which was disabled during the most recent
LLVM rebase. RAUW is only applicable to ReplaceableMetadata, which general
MDNodes (like those constructed for TBAA) are not. Fortunately,
replaceOperandWith() can be used for MDNodes with existing operands, which is
why we need to add an additional layer to the hierarchy.

Change-Id: I9f0def0eb109099f12e5292bc1f3a6deafd4aad9