Sanitize Module's target information.
We have always used a command line specified triple when instantiating
TargetMachine, rather than the triple from the Module.  Now sanitize
the Module by installing this same triple in the Module, and installing
the TargetMachine's data layout in the Module; this way, the Module and
the TargetMachine will be consistent.

Change-Id: I07480ba2c6effbabe15f72755d9d5792ed494fb0