Fix static object construction.
allocPointersExposed() maintains a static vector to hold a table of
function names. Unintentionally, the size of the table could grow with
each function call because the entries were being added by appending to
the existing object.

This can be fixed by using an initialization list instead of a
sequence of appends. Also, mark the table as const to ensure the code
doesn't try to modify it.

Change-Id: I0da0f99928e1614c945befff550b039d4f21e171