Fix a bug where kernels could modify the input allocation.
Bug: 22236496

A large struct in an input Allocation, passed by reference in AArch64,
was not being properly copied into a temporary location as the calling
convention requires. As a result, the kernel could have modified the
input allocation if it made any modifications to the argument.

Fix by copying the struct to a stack location before passing it to the
kernel.

Change-Id: I5089f9e5e27cd0d4ff249cbdc21f0c6ae186f189
(cherry picked from commit d96c9fae6bcbaa7d8bee0dab2d75beb8400248ff)