Fix rs_kernel_context_t type mismatch in call to kernel.
The kernel function context argument (if any) is a pointer to an
opaque user-visible type (rs_kernel_context_t) that differs from
the RsExpandKernelDriverInfoPfx type used in the wrapper
function (although the two types represent the same thing).
Therefore, we must introduce a pointer cast when generating a
call to the kernel function.

Bug: 21812992
Change-Id: I7680676a159cf052e2761b4f670f67891c3d6c1b