Change libbcc so that it compiles as an LLVM loadable module (try 2).
The main change is not requiring libLLVM in the makefile for libbcc.
With this change, libbcc can be used as an LLVM loadable module, for
instance as in "opt -load libbcc.so". This is helpful for debugging
individual bcc passes. Note that we still must link directly against
libLLVM on Windows due to the fact that DLLs can't access unresolved
symbols.

Revert "Revert "Change libbcc so that it compiles as an LLVM loadable module.""

This reverts commit a2bbdcdec92f076a38e75bbd81542fc49da19862.

Change-Id: I6729c2fed8542579ad76fcc533db091138f5575e