Fix alias metadata issues.
Bug: 22804168

(1) Remove the used of alias.scope metadata. This was not being used
correctly and had no effect.

(2) Fix the use of TBAA. Ensure that the temporary alloca used for
passing pointers to the kernel (if the kernel takes a pointer
parameter) does not get annotated with the RenderScript TBAA, because
this can conflict with existing TBAA metadata in the kernel.

Change-Id: I27018f809a33f213c6d4cc720b18b844f28b2d6f