Restore missing min and max API.
min(float2/3/4, float) and max(float2/3/4, float) were defined in our
headers in K.  In L, they were removed by mistake.  This continued to
work however because a float would be auto converted by the compiler to
a float4.

In M, we started validating for unexpected external references, which
_Z3minDv4_ff is.  This can prevent programs compiled with K from running
on M.

This CL fixes the whitelist.

Bug:23389429

Change-Id: I21f59521b3744d0c5b6fd75dbf1810493e9c1d84